+++ vehicle/src/Vehicle/Compile/Warning.hs
@@ -7,6 +7,7 @@ import Vehicle.Prelude
 import Vehicle.Resource (ExternalResource)
 import Vehicle.Syntax.AST
 import Vehicle.Verify.Core
+import Vehicle.Verify.QueryFormat.Core
 
 -- TODO we should really integrate these directly into `MonadLogger` and
 -- then implement caching to avoid printing multiple equal warnings out twice
@@ -16,6 +17,7 @@ data CompileWarning
   = UnusedResource ExternalResource (Set Name)
   | ResortingtoFMElimination Name (Set MixedVariable)
   | TrivialProperty PropertyAddress Bool
+  | UnsoundStrictOrderConversion Name QueryFormatID
 
 instance Pretty CompileWarning where
   pretty = \case
@@ -44,3 +46,22 @@ instance Pretty CompileWarning where
         <+> pretty status
         <+> "without needing to call the verifier. This usually indicates a fault with either the"
         <+> "specification or any external datasets used."
+    UnsoundStrictOrderConversion propertyName queryFormatID ->
+      "While compiling property"
+        <+> quotePretty propertyName
+        <+> "at least one of the generated SAT problems generated was found"
+        <+> "to contain a strict inequality (i.e. constraints of the form 'x < y')."
+        <+> "Unfortunately the"
+        <+> pretty queryFormatID
+        <+> "only supports non-strict"
+        <+> "inequalities (i.e. constraints of the form 'x <= y')."
+        <> line
+        <> line
+        <> "In order to provide support, Vehicle has automatically converted the"
+          <+> "strict inequalities to non-strict inequalites."
+          <+> "This is not sound, but errors will be at most the floating point epsilon"
+          <+> "used by the verifier, which is usually very small (e.g. 1e-9)."
+          <+> "However, this may lead to unexpected behaviour (e.g. loss of the law of excluded middle)."
+        <> line
+        <> line
+        <> "See https://github.com/vehicle-lang/vehicle/issues/74 for further details."