+++ vehicle/src/Vehicle/Verify/QueryFormat/Marabou.hs
@@ -3,7 +3,7 @@ module Vehicle.Verify.QueryFormat.Marabou
   )
 where
 
-import Control.Monad (forM)
+import Control.Monad.Writer
 import Data.Bifunctor (Bifunctor (..))
 import Data.List.NonEmpty (NonEmpty (..))
 import Data.Map (Map)
@@ -36,7 +36,10 @@ marabouQueryFormat =
 -- | Compiles an expression representing a single Marabou query. The expression
 -- passed should only have conjunctions and existential quantifiers at the boolean
 -- level.
-compileMarabouQuery :: (MonadLogger m) => CLSTProblem -> m QueryText
+compileMarabouQuery ::
+  (MonadLogger m, MonadWriter UnsoundStrictOrderConversion m) =>
+  CLSTProblem ->
+  m QueryText
 compileMarabouQuery (CLSTProblem variables assertions) = do
   let variableNames = sequentialNetworkVariableNaming "x" "y" variables
   let variableNamesMap = Map.fromList (zip variables variableNames)
@@ -45,7 +48,7 @@ compileMarabouQuery (CLSTProblem variables assertions) = do
   return $ layoutAsText assertionsDoc
 
 compileAssertion ::
-  (MonadLogger m) =>
+  (MonadLogger m, MonadWriter UnsoundStrictOrderConversion m) =>
   Map NetworkVariable Name ->
   Assertion NetworkVariable ->
   m (Doc a)
@@ -63,20 +66,24 @@ compileAssertion varNames assertion = do
           _ -> (coeffVars, rel, constant, True)
 
   let compiledLHS = hsep (fmap (compileVar multipleVariables) coeffVars')
-  let compiledRel = compileRel rel'
+  compiledRel <- compileRel rel'
   let compiledRHS = prettyRationalAsFloat constant'
 
   return $ compiledLHS <+> compiledRel <+> compiledRHS
 
-compileRel :: Either () OrderOp -> Doc a
+compileRel :: (MonadWriter UnsoundStrictOrderConversion m) => Either () OrderOp -> m (Doc a)
 compileRel = \case
-  Left () -> "="
-  Right Le -> "<="
-  Right Ge -> ">="
+  Left () -> return "="
+  Right Le -> return "<="
+  Right Ge -> return ">="
   -- Suboptimal. Marabou doesn't currently support strict inequalities.
   -- See https://github.com/vehicle-lang/vehicle/issues/74 for details.
-  Right Lt -> "<="
-  Right Gt -> ">="
+  Right Lt -> do
+    tell (Any True)
+    return "<="
+  Right Gt -> do
+    tell (Any True)
+    return ">="
 
 compileVar :: Bool -> (Rational, Name) -> Doc a
 compileVar False (1, var) = pretty var